+++ src/Helper/InstallationMethod.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Castor\Helper;
+
+enum InstallationMethod: string
+{
+    case Phar = 'phar';
+    case Static = 'static';
+    case ComposerGlobal = 'composer global';
+    case Composer = 'composer';
+    case Source = 'source';
+}